Start Notifier after availability update to ensure nodes finished their self repair
