+++ lib/archethic/self_repair/supervisor.ex
@@ -3,7 +3,7 @@ defmodule Archethic.SelfRepair.Supervisor do
 
   use Supervisor
 
-  alias Archethic.SelfRepair.{Notifier, Scheduler}
+  alias Archethic.SelfRepair.Scheduler
   alias Archethic.Utils
 
   def start_link(arg) do
@@ -13,8 +13,8 @@ defmodule Archethic.SelfRepair.Supervisor do
   def init(_arg) do
     children = [
       {Scheduler, Application.get_env(:archethic, Scheduler)},
-      {DynamicSupervisor, strategy: :one_for_one, name: Notifier.RepairSupervisor},
-      {Registry, name: Notifier.RepairRegistry, keys: :unique, partitions: 1}
+      {DynamicSupervisor, strategy: :one_for_one, name: Archethic.SelfRepair.NotifierSupervisor},
+      {Registry, name: Archethic.SelfRepair.Notifier.RepairRegistry, keys: :unique, partitions: 1}
     ]
 
     Supervisor.init(Utils.configurable_children(children), strategy: :one_for_one)
