+++ lib/archethic/self_repair/sync.ex
@@ -271,7 +271,7 @@ defmodule Archethic.SelfRepair.Sync do
     if :persistent_term.get(:archethic_up, nil) == :up do
       new_available_nodes = P2P.authorized_and_available_nodes(availability_update)
 
-      Notifier.start(previous_available_nodes, new_available_nodes)
+      Notifier.start_notifier(previous_available_nodes, new_available_nodes, availability_update)
     end
 
     update_statistics(summary_time, transaction_summaries)
